Code Refactoring Added PageState and PageStateNotifier which able to accept a Generic type <T> so that we are able to reuse it whenever the listview requires pagination.
Feature
Added routing configuration with go_router package. The package is come from flutter.dev which provide a comprehensive API. One of it is what we are looking for which is nested navigation. The app has bottom navigation bar which required nested navigation feature that does not override the base of the screen (bottom navigation remain at the bottom of the screen).